refactoring ad change db
