+++ LOCS/backend/static/JS/scripts/DBinit.js
@@ -1,0 +1,8 @@
+const config = require('../../../DBsetings.json');
+let pgp = require("pg-promise")( /*options*/ );
+let db = pgp(config);
+
+module.exports = {
+    pgp,
+    db
+};
