+++ LOCS/backend/static/JS/scripts/DataBase.js
@@ -1,6 +1,6 @@
-const config = require('../../../DBsetings.json');
-let pgp = require("pg-promise")( /*options*/ );
-let db = pgp(config);
+const intiDB = require('./DBinit.js');
+let db = intiDB.db;
+
 
 //проверка, что такое мыло есть
 const CheckUser = (mail) => {
@@ -433,32 +433,32 @@ let returnTokenAcceptMail = (id) => {
 
 
 
-////
-//Добавить район
-let addDistrict = (title, id_city) => {
-    return new Promise((resolve, reject) => {
-        db.result('Call AddDistrict($1, $2);', [title, id_city])
-            .then(function(data) {
-                resolve(true);
-            }).catch(function() {
-                reject("ERROR BD: AddDistrict");
-                return;
-            });
-    })
-}
+// ////
+// //Добавить район
+// let addDistrict = (title, id_city) => {
+//     return new Promise((resolve, reject) => {
+//         db.result('Call AddDistrict($1, $2);', [title, id_city])
+//             .then(function(data) {
+//                 resolve(true);
+//             }).catch(function() {
+//                 reject("ERROR BD: AddDistrict");
+//                 return;
+//             });
+//     })
+// }
 
-//Добавить адрес
-let addAddress = (street, house, latitude, longitude, idDistrict) => {
-    return new Promise((resolve, reject) => {
-        db.result('Call AddAddress($1, $2,$3,$4,$5);', [street, house, latitude, longitude, idDistrict])
-            .then(function(data) {
-                resolve(true);
-            }).catch(function() {
-                reject("ERROR BD: AddAddress");
-                return;
-            });
-    })
-}
+// //Добавить адрес
+// let addAddress = (street, house, latitude, longitude, idDistrict) => {
+//     return new Promise((resolve, reject) => {
+//         db.result('Call AddAddress($1, $2,$3,$4,$5);', [street, house, latitude, longitude, idDistrict])
+//             .then(function(data) {
+//                 resolve(true);
+//             }).catch(function() {
+//                 reject("ERROR BD: AddAddress");
+//                 return;
+//             });
+//     })
+// }
 
 
 ///////////////////
@@ -534,34 +534,34 @@ let tagById = (id) => {
 
 };
 
-// удалить тег по id
-let deleteTag = (id) => {
-    return new Promise((resolve, reject) => {
-        db.result('call deleteTag($1);', [id])
-            .then(function(data) {
-                resolve(data);
-            }).catch(function(e) {
-                console.log(e);
-                reject("ERROR BD: deleteTag");
-                return;
-            });
-    })
+// // удалить тег по id
+// let deleteTag = (id) => {
+//     return new Promise((resolve, reject) => {
+//         db.result('call deleteTag($1);', [id])
+//             .then(function(data) {
+//                 resolve(data);
+//             }).catch(function(e) {
+//                 console.log(e);
+//                 reject("ERROR BD: deleteTag");
+//                 return;
+//             });
+//     })
 
-};
+// };
 
-// подтв. тег 
-let acceptTag = (id) => {
-    return new Promise((resolve, reject) => {
-        db.result('call acceptTag($1);', [id])
-            .then(function(data) {
-                resolve(data);
-            }).catch(function(e) {
-                console.log(e);
-                reject("ERROR BD: acceptTag");
-                return;
-            });
-    })
-};
+// // подтв. тег 
+// let acceptTag = (id) => {
+//     return new Promise((resolve, reject) => {
+//         db.result('call acceptTag($1);', [id])
+//             .then(function(data) {
+//                 resolve(data);
+//             }).catch(function(e) {
+//                 console.log(e);
+//                 reject("ERROR BD: acceptTag");
+//                 return;
+//             });
+//     })
+// };
 
 // добавить тег 
 let addTag = (title) => {
@@ -1108,11 +1108,11 @@ module.exports = {
     'returnTokenAcceptMail': returnTokenAcceptMail, //вернуть токен по id для подтв аккаунт
 
     //для админки 
-    'addDistrict': addDistrict, //Добавить район
-    'addAddress': addAddress, //Добавить адрес
+    //'addDistrict': addDistrict, //Добавить район
+    //'addAddress': addAddress, //Добавить адрес
     'addTag': addTag, //добавить тег
-    'acceptTag': acceptTag, // подтв. тег
-    'deleteTag': deleteTag, //удалить тег
+    // 'acceptTag': acceptTag, // подтв. тег
+    //  'deleteTag': deleteTag, //удалить тег
 
     //события 
     'EventTags': EventTags, //Теги по Id евента
