+++ query scripts and shemas/procedures/organizer/get id adress.txt
@@ -1,3 +1,16 @@
+
+--Функция, которая возвращает id адреса
+CREATE OR REPLACE FUNCTION getAddress(street_ text, house_ text) 
+RETURNS table (j json) AS
+$$
+BEGIN
+    RETURN QUERY  SELECT to_json(a) FROM (select id from address where street like street_ and house like house_ and deleted = false)a;
+END;
+$$ LANGUAGE plpgsql;
+
+
+
+//старая
 --Функция, которая возвращает id адреса
 CREATE OR REPLACE FUNCTION getAddress(street_ text, house_ text) 
 RETURNS table (j json) AS
