+++ query scripts and shemas/procedures/organizer/search adress.txt
@@ -1,3 +1,14 @@
+
+--поиск адреса
+CREATE OR REPLACE FUNCTION searchAddress(word_ text) 
+RETURNS table (j json) AS
+$$
+BEGIN
+    RETURN QUERY  SELECT to_json(a) FROM (select id,street,house from address where deleted = false and (street ilike word_ or house ilike word_) LIMIT 15)a;
+END;
+$$ LANGUAGE plpgsql;
+
+//старая версия
 --поиск адреса
 CREATE OR REPLACE FUNCTION searchAddress(word_ text) 
 RETURNS table (j json) AS
